LPD-98985 Extract the broken links query into its own collaborator
